Get MatchServer to run on Linux
This commit mostly just removes MSVC/Windows-specific things like DWORD, strcpy_s, etc., along with hiding a bunch of things behind #ifdef _WIN32.

Most of RealSpace2 is not built at all on Linux (since it uses Direct3D9 directly all the time), it just builds a few things like RAnimation, RBspObject and their dependencies, since the code for server-based hit registration in MatchServer uses those.

It also replaces all of the many instances of unsigned long int in MMatchTransDataTypes.h with u32, since the size of long is not portable at all: It's 32-bits on MSVC like we expect, but 64-bits on gcc and clang on a 64-bit arch. Additionally, MMatchObjCache's destructor is made non-virtual, making it a standard-layout class again, which is necessary since the MC_MATCH_OBJECT_CACHE (foolishly) directly sends this type, and so the size of the blob would end up depending on the size of the vtable pointer. Since I think it's used in replays somewhere, a dummy u32 is added to the top to stand in for the missing vtable.

MRealCPNet in MServer is replaced with the new type NetIO, which under Windows is just a thin wrapper for MRealCPNet, but on Linux is implemented in terms of the asio library.

Known bugs: NetIO::Create can set up more than thread for running handlers, but it only seems to work with one, which is the current setting.

+++ CSCommon/Include/MCommandCommunicator.h
@@ -25,7 +25,7 @@ protected:
 	MPacketCrypter			m_PacketCrypter;
 
 	MCommandCommunicator*	m_pDirectConnection;
-	u32						m_dwUserContext;
+	uintptr_t				m_dwUserContext;
 
 	char					m_szIP[128];
 	int						m_nPort;
@@ -44,8 +44,8 @@ public:
 
 	MCommandCommunicator* GetDirectConnection()			{ return m_pDirectConnection; }
 	void SetDirectConnection(MCommandCommunicator* pDC)	{ m_pDirectConnection = pDC; }
-	u32 GetUserContext() const							{ return m_dwUserContext; }
-	void SetUserContext(u32 dwContext)					{ m_dwUserContext = dwContext; }
+	uintptr_t GetUserContext() const					{ return m_dwUserContext; }
+	void SetUserContext(uintptr_t dwContext)			{ m_dwUserContext = dwContext; }
 
 	auto* GetIPString() const { return m_szIP; }
 	u32 GetIP() const		{ return m_dwIP; }
